--- ConsoleApp7/ConsoleApp7.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.27703.2042
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ConsoleApp7", "ConsoleApp7.csproj", "{9BC039D6-60E6-45AA-B3FE-7508960F193A}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{9BC039D6-60E6-45AA-B3FE-7508960F193A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{9BC039D6-60E6-45AA-B3FE-7508960F193A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{9BC039D6-60E6-45AA-B3FE-7508960F193A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{9BC039D6-60E6-45AA-B3FE-7508960F193A}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {A2999C1A-673F-47DC-82AF-8F0F5051A404}
-	EndGlobalSection
-EndGlobal
