+++ src/ExpressionMapper/ExpressionMapper.csproj
@@ -1,0 +1,10 @@
+﻿<Project Sdk="Microsoft.NET.SDK">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard2.0</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="System.ComponentModel.Annotations" Version="4.5.0" />
+  </ItemGroup>
+</Project>
